celebritySearch/pyproject.toml
[project]
name = "celebritysearch"
version = "0.1.0"
description = "Search any celebrity"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-openai>=0.2.12",
    "langchain>=0.3.11",
    "langchain-community>=0.3.11",
    "openai>=1.57.4",
    "python-dotenv>=1.0.1",
    "streamlit>=1.41.1",
]
